Move BindInputs to typebind.go
